requirements_env.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
azure-common==1.1.28
azure-core==1.38.2
azure-identity==1.25.2
azure-mgmt-authorization==4.0.0
azure-mgmt-compute==37.2.0
azure-mgmt-containerservice==40.2.0
azure-mgmt-core==1.6.0
azure-mgmt-cosmosdb==9.9.0
azure-mgmt-keyvault==10.3.0
azure-mgmt-network==30.2.0
azure-mgmt-resource==25.0.0
azure-mgmt-sql==3.0.1
azure-mgmt-storage==21.1.0
azure-storage-blob==12.28.0
blinker==1.9.0
boto3==1.42.62
botocore==1.42.62
cachetools==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
idna==3.11
isodate==0.7.2
Jinja2==3.1.6
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msal==1.35.1
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.7.1
narwhals==2.17.0
neo4j==6.1.0
networkx==3.6.1
numpy==1.26.4
oauthlib==3.3.1
packaging==23.2
pandas==2.2.1
pillow==10.4.0
plotly==5.19.0
propcache==0.4.1
protobuf==4.25.8
pyarrow==23.0.1
pycparser==3.0
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.11.0
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.2.1
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.30.0
s3transfer==0.16.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.32.2
tenacity==8.5.0
toml==0.10.2
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==4.0.0
yarl==1.23.0
